First cut of refactoring
